plot functions are moved to separated module
