+++ config/inference.yaml
@@ -1,11 +1,11 @@
 line_detection:
   model_path: ..\models\lines_obb_m_best.pt
-  min_conf: 0.55
+  min_conf: 0.4
   intersection_resolver: MeanResolver
 
 word_detection:
   model_path: ..\models\word_detect_m_best.pt
   min_conf: 0.4
-  intersection_resolver: ByConfResolver
+  intersection_resolver: UnionResolver
 
 prediction_dir: ..\predictions
